tests: Move healthcheck helper class to test_utils
